botones ver ingresos terminados

+++ src/main/java/com/AppDinero2/App/repository/IngresosRepository.java
@@ -3,5 +3,5 @@ package com.AppDinero2.App.repository;
 import com.AppDinero2.App.entity.Ingreso;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface IngresosRepository extends JpaRepository<Ingreso,Long> {
+public interface IngresosRepository extends JpaRepository<Ingreso,Integer> {
 }
